[JENKINS-35797] Refactor implementation
* Move create() to BlueRunContainer.
* Return BlueQueueItem on create()
