+++ blueocean-rest-impl/src/main/java/io/jenkins/blueocean/service/embedded/rest/RunContainerImpl.java
@@ -1,14 +1,23 @@
 package io.jenkins.blueocean.service.embedded.rest;
 
+import com.google.common.base.Predicate;
+import hudson.model.Cause;
+import hudson.model.CauseAction;
 import hudson.model.Job;
+import hudson.model.Queue;
+import hudson.model.queue.ScheduleResult;
 import hudson.util.RunList;
 import io.jenkins.blueocean.commons.ServiceException;
 import io.jenkins.blueocean.rest.hal.Link;
 import io.jenkins.blueocean.rest.model.BluePipeline;
+import io.jenkins.blueocean.rest.model.BlueQueueItem;
 import io.jenkins.blueocean.rest.model.BlueRun;
 import io.jenkins.blueocean.rest.model.BlueRunContainer;
+import jenkins.model.Jenkins;
+import org.jenkinsci.plugins.github.util.FluentIterableWrapper;
 
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import java.util.Iterator;
 
 /**
@@ -60,5 +69,43 @@ public class RunContainerImpl extends BlueRunContainer {
     @Override
     public BluePipeline getPipeline(String name) {
         return pipeline;
+    }
+
+    /**
+     * Schedules a build. If build already exists in the queue and the pipeline does not
+     * support running multiple builds at the same time, return a reference to the existing
+     * build.
+     *
+     * @return Qeueu item.
+     */
+    @Override
+    public BlueQueueItem create() {
+        if (job instanceof Queue.Task) {
+            ScheduleResult scheduleResult = Jenkins.getInstance()
+                .getQueue()
+                .schedule2((Queue.Task)job, 0, new CauseAction(new Cause.UserIdCause()));
+
+            if(scheduleResult.isAccepted()) {
+                final Queue.Item item = scheduleResult.getItem();
+
+                BlueQueueItem queueItem = FluentIterableWrapper.from(QueueContainerImpl.getQueuedItems(job, pipeline))
+                    .firstMatch(new Predicate<BlueQueueItem>() {
+                        @Override
+                        public boolean apply(@Nullable BlueQueueItem input) {
+                            return input.getId().equalsIgnoreCase(Long.toString(item.getId()));
+                        }
+                    }).orNull();
+
+                if (queueItem == null) {
+                    throw new ServiceException.UnexpectedErrorException("The queue item does not exist in the queue");
+                } else {
+                    return queueItem;
+                }
+            } else {
+                throw new ServiceException.UnexpectedErrorException("Queue item request was not accepted");
+            }
+        } else {
+            throw new ServiceException.NotImplementedException("This pipeline type does not support being queued.");
+        }
     }
 }
