+++ blueocean-rest/src/main/java/io/jenkins/blueocean/rest/model/BlueRunContainer.java
@@ -1,5 +1,9 @@
 package io.jenkins.blueocean.rest.model;
 
+import io.jenkins.blueocean.commons.stapler.TreeResponse;
+import org.kohsuke.stapler.WebMethod;
+import org.kohsuke.stapler.verb.POST;
+
 /**
  * BlueRun API
  *
@@ -12,4 +16,9 @@ public abstract class BlueRunContainer extends Container<BlueRun> {
      * @return pipeline with the given name as parameter
      */
     public abstract BluePipeline getPipeline(String name);
+
+    @POST
+    @WebMethod(name = "")
+    @TreeResponse
+    public abstract BlueQueueItem create();
 }
